Moves options to src/lib/
